Added cursor-pointer to all buttons and clickable object

+++ db/microgreens_db_procedures.sql
@@ -1113,8 +1113,9 @@ BEGIN
 		WHERE order_id = order_id_p
 		AND product_id = product_id_p;
 
-	DELETE FROM customer_order
-		WHERE order_id = order_id_p;
+	IF NOT EXISTS ( SELECT 1 FROM contains WHERE order_id = order_id_p) THEN
+		DELETE FROM customer_order WHERE order_id = order_id_p;
+	END IF;
 END //
 DELIMITER ;
